[REFECTOR] #41 - empty, exist fragment 하나로 합침
